Capstone_2/requirements.txt
numpy
Pillow
tflite_runtime
requests